pyproject.toml
[project]
name = "hnc-poc"
version = "0.1.0"
description = "Mapillary AOI to TRIBE v2 brain-encoding to GeoParquet 2.0"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "CC-BY-NC-4.0" }
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "duckdb>=1.5",
    "pyarrow>=15",
    "numpy>=1.26",
    "tqdm",
    "typer>=0.12",
    "Pillow>=10",
    "imageio[ffmpeg]>=2.34",
    "huggingface_hub>=0.24",
]

[project.optional-dependencies]
tribe = [
    "tribev2 @ git+https://github.com/facebookresearch/tribev2.git",
    "mne>=1.7",
    "nilearn>=0.10",
    "torch>=2.4",
    "torchvision>=0.19",
]
dev = ["pytest", "ruff", "mypy"]

[project.scripts]
hnc-run = "src.pipeline:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "C4",   # comprehensions
]
ignore = [
    "E501",  # line too long, ruff format handles this
]

[tool.ruff.lint.flake8-bugbear]
# typer.Option / typer.Argument in defaults is the idiomatic pattern, allowlist them.
extend-immutable-calls = ["typer.Option", "typer.Argument"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true   # tribev2, mne, nilearn lack type stubs

[tool.uv]
# uv reads dependencies from [project], no extra config needed for the basic flow
